Extract LDP copy logic to ldp-copy module, per Nicola's request
