+++ lib/handlers/copy.js
@@ -1,23 +1,9 @@
 module.exports = handler
 
-const utils = require('../utils')
-const mkdirp = require('fs-extra').mkdirp
-const fs = require('fs')
-const request = require('request')
-const path = require('path')
 const debug = require('../debug')
 const error = require('../http-error')
-
-/**
- * Cleans up a file write stream (ends stream, deletes the file).
- * @method cleanupFileStream
- * @param stream {WriteStream}
- */
-function cleanupFileStream (stream) {
-  let streamPath = stream.path
-  stream.destroy()
-  fs.unlinkSync(streamPath)
-}
+const ldpCopy = require('../ldp-copy')
+const utils = require('../utils')
 
 /**
  * Handles HTTP COPY requests to import a given resource (specified in the
@@ -36,41 +22,15 @@ function handler (req, res, next) {
   const copyFromUrl = serverRoot + copyFrom
   const copyTo = res.locals.path || req.path
   const copyToPath = utils.reqToPath(req)
-
-  mkdirp(path.dirname(copyToPath), function (err) {
+  ldpCopy(copyToPath, copyFromUrl, function (err) {
     if (err) {
-      debug.handlers('COPY -- Error creating destination directory: ' + err)
-      return next(error(500,
-        'Failed to create the path to the destination resource'))
+      let statusCode = err.statusCode || 500
+      let errorMessage = err.statusMessage || err.message
+      debug.handlers('Error with COPY request:' + errorMessage)
+      return next(error(statusCode, errorMessage))
     }
-    var destinationStream = fs.createWriteStream(copyToPath)
-      .on('error', function (err) {
-        cleanupFileStream(this)
-        return next(error(500, 'Error writing data: ' + err))
-      })
-      .on('finish', function () {
-        debug.handlers('COPY -- Wrote data to: ' + copyToPath)
-        res.set('Location', copyTo)
-        res.sendStatus(201)
-        next()
-      })
-    request.get(copyFromUrl)
-      .on('error', function (err) {
-        debug.handlers('COPY -- Error requesting source file: ' + err)
-        this.end()
-        cleanupFileStream(destinationStream)
-        return next(error(500, 'Error writing data: ' + err))
-      })
-      .on('response', function (response) {
-        if (response.statusCode !== 200) {
-          debug.handlers('COPY -- HTTP error reading source file: ' +
-            response.statusMessage)
-          this.end()
-          cleanupFileStream(destinationStream)
-          return next(error(response.statusCode,
-            'Error reading source file: ' + response.statusMessage))
-        }
-      })
-      .pipe(destinationStream)
+    res.set('Location', copyTo)
+    res.sendStatus(201)
+    return next()
   })
 }
